+++ CHANGELOG.md
@@ -4,6 +4,8 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
+### Added
+- Add DefaultHttpCaller backed by java native HttpClient to perform simple get API invocation
 
 # [17.105.0-M3] - 2025-02-04
 ### Added